Remove HahtiTrust plugin
I suggest that we remove the HathiTrust option from Primo Studio. This package is based on an unmaintained private fork that does not work with VE. Given that most of the community is now on VE, the presence of this forked package seems to cause nothing but confusion. More info: https://github.com/UMNLibraries/primo-explore-hathitrust-availability/issues/13.